+++ R/method-as-ultrametric.R
@@ -12,22 +12,31 @@ as.ultrametric <- function(tree, ...){
 ## reference 
 ## https://github.com/PuttickMacroevolution/MCMCtreeR/blob/master/R/readMCMCTree.R
 as.ultrametric.phylo <- function(tree, ...){
-    outer <- tree$edge[, 2]
-    inner <- tree$edge[, 1]
+    edge <- tree$edge
+    outer <- edge[, 2]
+    inner <- edge[, 1]
     ntip <- Ntip(tree)
-    totalPath <- c()
-    tipindx <- which(outer <= ntip)
-    for (i in tipindx) {
-        start <- i
-        end <- inner[start]
-        edgeTimes <- tree$edge.length[start]
-        while (end != inner[1]) {
-            start <- which(outer == end)
-            end <- inner[start]
-            edgeTimes <- c(edgeTimes, tree$edge.length[start])
+    root <- rootnode(tree)
+    children <- split(seq_along(outer), inner)
+    distance <- numeric(max(edge))
+    queue <- integer(max(edge))
+    queue[1] <- root
+    n_queue <- 1L
+    i <- 1L
+    while (i <= n_queue) {
+        edge_index <- children[[as.character(queue[i])]]
+        n_child <- length(edge_index)
+        if (n_child > 0) {
+            child <- outer[edge_index]
+            distance[child] <- distance[queue[i]] + tree$edge.length[edge_index]
+            idx <- n_queue + seq_len(n_child)
+            queue[idx] <- child
+            n_queue <- n_queue + n_child
         }
-        totalPath <- c(totalPath, sum(edgeTimes))
+        i <- i + 1L
     }
+    tipindx <- which(outer <= ntip)
+    totalPath <- distance[outer[tipindx]]
     addLength <- max(totalPath) - totalPath
     tree$edge.length[tipindx] <- tree$edge.length[tipindx] + addLength
     return (tree)